w3c-reports/12092014/gtk/runned-tests/dom/lists/DOMTokenList-stringifier-expected.txt


FAIL DOMTokenList stringifier assert_equals: String(classList) should compress whitespace expected "a b" but got "   a  a b"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/lists/DOMTokenList-stringifier.html:17:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/dom/lists/DOMTokenList-stringifier.html:11:5)
